pyproject.toml
[project]
name = "openbankapi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.117.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlmodel>=0.0.25",
    "uvicorn>=0.37.0",
]
